requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.7.2
httpie==2.4.0
httplib2==0.19.1
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
wrapt==1.12.1
